Create a citation of a Source record

Create a citation of a Source record

source_citation(
  gedcom,
  source,
  where = character(),
  event = character(),
  role = character(),
  entry_date = character(),
  source_text = character(),
  certainty = character(),
  notes = character(),
  multimedia_links = character()
)

Arguments

gedcom

A tidyged object.

source

A character string identifying the source. This can either be an xref or term(s) to match to a source title.

where

Specific location within the information referenced. For a published work, this could include the volume of a multi-volume work and the page number(s). For a newspaper, it could include a column number and page number. A census record might have an enumerating district, page number, line number, dwelling number, and family number. The data in this field should be in the form of a label and value pair, such as Label1: value, Label2: value, with each pair being separated by a comma. For example, Film: 1234567, Frame: 344, Line: 28.

event

A code that indicates the type of event which was responsible for the source entry being recorded. For example, if the entry was created to record a birth of a child, then the type would be BIRT regardless of the assertions made from that record, such as the mother's name or mother's birth date. This will allow a prioritised best view choice and a determination of the certainty associated with the source used in asserting the cited fact.

role

Indicates what role this person played in the event that is being cited in this context.

entry_date

A date_calendar(), date_period(), date_range(), or date_approximated() value giving the date that this data was entered into the original source document.

source_text

A verbatim copy of any description contained within the source. This indicates notes or text that are actually contained in the source document, not the submitter's opinion about the source.

certainty

An evaluation of the credibility of a piece of information, based upon its supporting evidence. Some systems use this feature to rank multiple conflicting opinions for display of most likely information first. It is not intended to eliminate the receiver's need to evaluate the evidence for themselves. Values allowed: "unreliable", "subjective", "secondary", "primary".

notes

A character vector of notes accompanying the citation. These could be xrefs to existing Note records.

multimedia_links

A character vector of Multimedia record xrefs accompanying this record.

Value

A tibble describing a source citation.