Add a family event to a Family Group record

Add a family event to a Family Group record

add_family_event(
  gedcom,
  event_type = character(),
  event_descriptor = character(),
  event_classification = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_phonetic = character(),
  phonetisation_method = character(),
  place_romanised = character(),
  romanisation_method = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  local_address_lines = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  multimedia_links = character(),
  xref = character(),
  update_date_changed = TRUE
)

add_family_event_annulment(...)

add_family_event_census(...)

add_family_event_divorce(...)

add_family_event_divorce_filed(...)

add_family_event_engagement(...)

add_family_event_marriage_banns(...)

add_family_event_marriage_contract(...)

add_family_event_relationship(...)

add_family_event_marriage_license(...)

add_family_event_marriage_settlement(...)

add_family_event_residence(...)

add_family_event_other(...)

Arguments

gedcom
A tidyged object.

event_type
The type of event. This should be automatically populated with the appropriate event function.

event_descriptor
A short description of the event.

event_classification
A descriptive word or phrase used to further classify this event. This should be used whenever the 'other' event is used (but can also be used with others).

husband_age_at_event
A character string that indicates the age in years, months, and days that the husband was at the time of the event. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 10d".

wife_age_at_event
A character string that indicates the age in years, months, and days that the wife was at the time of the event. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 10d".

event_notes
A character vector of notes accompanying the event. These could be xrefs to existing Note records.

event_date
A date_value() object giving the date of the event.

event_cause
Used in special cases to record the reasons which precipitated an event. Normally this will be used for a death event to show cause of death, such as might be listed on a death certificate.

place_name
The jurisdictional name of the place where the event took place. Jurisdictions are separated by commas, for example, "Cove, Cache, Utah, USA."

place_phonetic
A character vector of phonetic variations of the place name.

phonetisation_method
A character vector giving the method used in transforming the text to the corresponding phonetic variation. If this argument is used, it must be the same size as the place_phonetic argument.

place_romanised
A character vector of romanised variations of the place name.

romanisation_method
A character vector giving the method used in transforming the text to the corresponding romanised variation. If this argument is used, it must be the same size as the place_romanised argument.

place_latitude
The value specifying the latitudinal coordinate of the event place. The latitude coordinate is the direction North or South from the equator in degrees and fraction of degrees carried out to give the desired accuracy. For example: 18 degrees, 9 minutes, and 3.4 seconds North would be formatted as "N18.150944"

place_longitude
The value specifying the longitudinal coordinate of the event place. The longitude coordinate is Degrees and fraction of degrees east or west of the zero or base meridian coordinate. For example: 168 degrees, 9 minutes, and 3.4 seconds East would be formatted as "E168.150944".

place_notes
A character vector of notes accompanying the event place. These could be xrefs to existing Note records.

local_address_lines
The first line of the event address.

city
The city of the event address.

state
The state/county of the event address.

postal_code
The postal code of the event address.

country
The country of the event address.

phone_number
A character vector containing up to three phone numbers of the event address.

email
A character vector containing up to three email addresses of the event address.

fax
A character vector containing up to three fax numbers of the event address.

web_page
A character vector containing up to three web pages of the event address.

responsible_agency
The organisation, institution, corporation, person, or other entity that has responsibility for the event data.

religious_affiliation
A name of the religion with which this event was affiliated.

multimedia_links
A character vector of multimedia file references accompanying this event. These could be xrefs to existing Multimedia records.

xref
The xref of a record to act on if one is not activated (will override active record).

update_date_changed
Whether to add/update the change date for the record.

...
See arguments for main function. The attribute_type/event_type do not need to be populated.

Value

An updated tidyged object with an expanded Family group record including this event.