Add a Multimedia record to a tidyged object

Add a Multimedia record to a tidyged object

add_media(
  gedcom,
  file_reference,
  format,
  source_media = character(),
  title = character(),
  user_reference_numbers = character(),
  media_notes = character()
)

Arguments

gedcom
A tidyged object.

file_reference
A reference for the file, typically a filepath or URL. It is strongly recommended you do not use backslashes, only use forward slashes.

format
A string indicating the format of the multimedia file. See Details.

source_media
A code that indicates the type of material in which the referenced source is stored. See Details.

title
The title of the multimedia file.

user_reference_numbers
A unique user-defined number or text that the submitter uses to identify this record. You can supply more than one in a vector. You can also define a user reference type by using a named vector (e.g c(id1 = "123A", id2 = "456B")).

media_notes
A character vector of notes accompanying this Multimedia record. These could be xrefs to existing Note records.

Value

An updated tidyged object including the Multimedia record.

Details

The formats allowed are currently limited to:

"AAC", "AVI", "BMP", "ePub", "FLAC", "GIF", "JPEG", "MKV", "mobi", "MP3", "PCX", "PDF", "PNG", "TIFF", "WAV".

The source media must be one of:

audio, book, card, electronic, fiche, film, magazine, manuscript, map, newspaper, photo, tombstone, video.