Add a source repository citation to a Source record

This structure is used within a source record to point to a name and address record of the holder of the source document.

add_sour_repo_citation(
  gedcom,
  repository,
  call_number = character(),
  xref = character(),
  update_date_changed = TRUE
)

Arguments

gedcom
A tidyged object.

repository
A character string identifying the repository. This can either be an xref or term(s) to match to a repository name.

call_number
An identification or reference description used to file and retrieve items from the holdings of a repository.

xref
The xref of a record to act on if one is not activated (will override active record).

update_date_changed
Whether to add/update the change date for the record.

Value

An updated tidyged object with an expanded Source record including this repository citation.