Add a Source record to a tidyged object

Add a Source record to a tidyged object

add_source(
  gedcom,
  events_recorded = character(),
  date_period_covered = date_period(),
  jurisdiction = character(),
  responsible_agency = character(),
  originator = character(),
  title = character(),
  short_title = character(),
  publication_detail = character(),
  source_text = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  data_notes = character(),
  source_notes = character(),
  multimedia_links = character()
)

Arguments

gedcom
A tidyged object.

events_recorded
An enumeration of the different kinds of events that were recorded in this source. Each enumeration is separated by a comma. See the Gedcom 5.5.5 Specification for more details.

date_period_covered
A date_period() object indicating the time period of events covered by this source.

jurisdiction
A character string indicating the lowest level jurisdiction encompassing all places named in this source. See the Gedcom 5.5.5 Specification for more details.

responsible_agency
The organisation, institution, corporation, person, or other entity that has responsibility for the source data.

originator
The person, agency, or entity who created the record.

title
The title of the source work, record, or item and, when appropriate, the title of the larger work or series of which it is a part.

short_title
An abbreviated or shortened version of the title (if required).

publication_detail
When and where the source record was created. For published works, this includes information such as the city of publication, name of the publisher, and year of publication.

source_text
A verbatim copy of relevant text contained within the source. This indicates notes or text that are actually contained in the source document, not the submitter's opinion about the source.

user_reference_number
A unique user-defined number or text that the submitter uses to identify this record. You can supply more than one in a vector.

user_reference_type
A user-defined definition of the user_reference_number(s). If this parameter is used, there must be a reference type for every reference number defined.

data_notes
A character vector of notes associated with the data in this Source record. These could be xrefs to existing Note records.

source_notes
A character vector of notes accompanying this Source record. These could be xrefs to existing Note records.

multimedia_links
A character vector of multimedia file references accompanying this source. These could be xrefs to existing Multimedia records.

Value

An updated tidyged object including the Source record.