Get descriptions for records

Get descriptions for records

describe_records(gedcom, xrefs, short_desc = FALSE)

Arguments

gedcom

A tidyged object.

xrefs

A vector of record xrefs. Only unique records are used. Header and trailer records are ignored.

short_desc

Whether to return a shorter description.

Value

A vector of record descriptions.

Examples

describe_records(sample555, sample555$record)
#> [1] "Submitter @U1@, Reldon Poulson"                                                                                                                                                                    
#> [2] "Individual @I1@, Robert Eugene Williams, born 2 OCT 1822 in Weston, Madison, Connecticut, United States of America, died 14 APR 1905 in Stamford, Fairfield, Connecticut, United States of America"
#> [3] "Individual @I2@, Mary Ann Wilson, born BEF 1828 in Connecticut, United States of America"                                                                                                          
#> [4] "Individual @I3@, Joe Williams, child of Robert Eugene Williams and Mary Ann Wilson, born 11 JUN 1861 in Idaho Falls, Bonneville, Idaho, United States of America"                                  
#> [5] "Family @F1@, headed by Robert Eugene Williams and Mary Ann Wilson, and children: Joe Williams"                                                                                                     
#> [6] "Family @F2@, headed by Robert Eugene Williams, and children: Joe Williams"                                                                                                                         
#> [7] "Source @S1@, titled Madison County Birth, Death, and Marriage Records"                                                                                                                             
#> [8] "Repository @R1@, Family History Library"