Helper functions to locate record xrefs

These functions act as wrappers to the find_xref function to find one or more record xrefs.

find_indi_refn(gedcom, pattern, ignore_case = FALSE)

find_indi_name(gedcom, pattern, ignore_case = FALSE)

find_indi_name_all(gedcom, pattern, ignore_case = FALSE)

find_repo_refn(gedcom, pattern, ignore_case = FALSE)

find_repo_name(gedcom, pattern, ignore_case = FALSE)

find_repo_name_all(gedcom, pattern, ignore_case = FALSE)

find_note_refn(gedcom, pattern, ignore_case = FALSE)

find_note_text(gedcom, pattern, ignore_case = FALSE)

find_note_text_all(gedcom, pattern, ignore_case = FALSE)

find_media_refn(gedcom, pattern, ignore_case = FALSE)

find_media_fileref(gedcom, pattern, ignore_case = FALSE)

find_media_fileref_all(gedcom, pattern, ignore_case = FALSE)

find_sour_refn(gedcom, pattern, ignore_case = FALSE)

find_sour_titl(gedcom, pattern, ignore_case = FALSE)

find_sour_titl_all(gedcom, pattern, ignore_case = FALSE)

find_famg_refn(gedcom, pattern, ignore_case = FALSE)

Arguments

gedcom
A tidyged object.

pattern
The search pattern to use (regular expression).

ignore_case
Should case differences be ignored in the match?

Value

A character vector of xref(s).

Details

If you have your own specific use cases to identify records, it's easy to write your own wrapper. It's best to name your function find_recordtype_* and end it with _all if it can return multiple xrefs. If you provide more than one search pattern, you should also include the mode argument.