Create a base tidyged object

This function creates a minimal tidyged object with header and footer sections and a single submitter record.

gedcom(
  submitter_details = subm(),
  gedcom_description = character(),
  gedcom_copyright = character(),
  source_data_name = character(),
  source_data_date = date_exact(),
  source_data_copyright = character(),
  receiving_system = "gedcompendium",
  language = "English"
)

Arguments

submitter_details

Details of the submitter of the file (you?) using the subm() function. If no submitter name is provided, the username is used.

gedcom_description

A note to describe the contents of the file in terms of "ancestors or descendants of" so that the person receiving the data knows what genealogical information the file contains.

gedcom_copyright

A copyright statement needed to protect the copyrights of the submitter of this GEDCOM file.

source_data_name

The name of the electronic data source that was used to obtain the data in this file.

source_data_date

The date this source was created or published. Ensure you create this date with the date_exact() function.

source_data_copyright

A copyright statement required by the owner of data from which this information was obtained.

receiving_system

The name of the system expected to process the GEDCOM-compatible file.

language

The human language in which the data in the file is normally read or written.

Value

A minimal tidyged object.