Identify all supporting records for a set of records

This function gets all supporting records (and onwards dependencies) for a set of records. Supporting records are note, multimedia, source, and repository records, i.e. those providing supporting evidence and comments.

get_supporting_records(
  gedcom,
  xrefs,
  include_note = TRUE,
  include_media = TRUE,
  include_sour = TRUE,
  include_repo = TRUE
)

Arguments

gedcom
A tidyged object.

xrefs
The xrefs of records to get supporting records for.

include_note
Whether to include Note records.

include_media
Whether to include Multimedia records.

include_sour
Whether to include Source records.

include_repo
Whether to include Repository records.

Value

A character vector of supporting record xrefs.