Define a personal name's components

Define a personal name's components

name_pieces(
  prefix = character(),
  given = character(),
  nickname = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character(),
  notes = character()
)

Arguments

prefix

The name prefix, e.g. Cmdr.

given

The given name or earned name. Different given names are separated by a comma.

nickname

A descriptive or familiar name used in connection with one's proper name.

surname_prefix

Surname prefix or article used in a family name. For example in the name "de la Cruz", this value would be "de la".

surname

Surname or family name. Different surnames are separated by a comma.

suffix

Non-indexing name piece that appears after the given name and surname parts, e.g. Jr. Different name suffix parts are separated by a comma.

notes

A character vector of notes accompanying this name. These could be xrefs to existing Note records.

Value

A tibble describing a personal name's components.