Define a place associated with a fact

Define a place associated with a fact

place(
  name = character(),
  phonetic_var = character(),
  phonetisation_method = character(),
  romanised_var = character(),
  romanisation_method = character(),
  latitude = character(),
  longitude = character(),
  notes = character()
)

Arguments

name
The jurisdictional name of the place. Jurisdictions are separated by commas, for example, "Cove, Cache, Utah, USA."

phonetic_var
A character vector of phonetic variations of the place name.

phonetisation_method
A character vector giving the method used in transforming the text to the corresponding phonetic variation. If this argument is used, it must be the same size as the phonetic_var argument.

romanised_var
A character vector of romanised variations of the place name.

romanisation_method
A character vector giving the method used in transforming the text to the corresponding romanised variation. If this argument is used, it must be the same size as the romanised_var argument.

latitude
The value specifying the latitudinal coordinate of the place. The latitude coordinate is the direction North or South from the equator in degrees and fraction of degrees carried out to give the desired accuracy. For example: 18 degrees, 9 minutes, and 3.4 seconds North would be formatted as "N18.150944"

longitude
The value specifying the longitudinal coordinate of the place. The longitude coordinate is Degrees and fraction of degrees east or west of the zero or base meridian coordinate. For example: 168 degrees, 9 minutes, and 3.4 seconds East would be formatted as "E168.150944".

notes
A character vector of notes accompanying the place. These could be xrefs to existing Note records.

Value

A tibble describing a place.