Add a family event to a Family Group record

Add a family event to a Family Group record

add_famg_event(
  gedcom,
  type,
  descriptor = "",
  classification = character(),
  date = character(),
  husband_age = character(),
  wife_age = character(),
  cause = character(),
  event_place = place(),
  event_address = address(),
  notes = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  multimedia_links = character(),
  xref = character(),
  update_date_changed = TRUE
)

Arguments

gedcom
A tidyged object.

type
A (case-insensitive) three-letter code giving the type of event. See Details.

descriptor
A short description of an 'other' event.

classification
A descriptive word or phrase used to further classify this event. This should be used whenever the 'other' event is used (but can also be used with others). Recommended values for non-marriage relationships can be found in Details.

date
A date_calendar(), date_approximated(), date_period(), or date_range() object giving the timing of the event..

husband_age
A character string that indicates the age in years, months, and days that the husband was at the time of the event. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 10d".

wife_age
A character string that indicates the age in years, months, and days that the wife was at the time of the event. Any combination of these is permitted. Any labels must come after their corresponding number, for example; "4y 8m 10d".

cause
Used in special cases to record the reasons which precipitated an event.

event_place
A place() object giving the place associated with this event.

event_address
An address() object giving the address associated with this event.

notes
A character vector of notes accompanying the event. These could be xrefs to existing Note records.

responsible_agency
The organisation, institution, corporation, person, or other entity that has responsibility for the event.

religious_affiliation
A name of the religion with which this event was affiliated.

multimedia_links
A character vector of Multimedia record xrefs accompanying this record.

xref
The xref of a record to act on if one is not activated (will override active record).

update_date_changed
Whether to add/update the change date for the record.

Value

An updated tidyged object with an expanded Family group record including this event.

Details

The three-letter codes used for the type parameter are:

ann(ulment), cen(sus), div(orce), dif (divorce filed), eng(agement), mab (marriage banns), mac (marriage contract), mal (marriage license), rel(ationship), mas (marriage settlement), res(idence).

Alternatively eve (for any other event).

Example classifications of non-marriage relationships are:

not married, civil, living together, living apart together. See page 59 of the GEDCOM Specification for more examples.